+++ packages/form/src/components/List/ListContainer.tsx
@@ -1,7 +1,7 @@
 ﻿import { PlusOutlined } from '@ant-design/icons';
 import { nanoid, runFunction } from '@ant-design/pro-utils';
 import { Button } from 'antd';
-import { omit } from 'lodash';
+import omit from 'omit.js';
 import { useMemo, useRef, useState } from 'react';
 import type { ProFormListItemProps } from './ListItem';
 import { ProFormListItem } from './ListItem';
@@ -82,7 +82,7 @@ const ProFormListContainer: React.FC<ProFormListItemProps> = (props) => {
           let index = uuidFields.length;
           // 如果是顶部，加到第一个，如果不是，为空就是最后一个
           if (position === 'top') index = 0;
-          await wrapperAction.add(runFunction(creatorRecord), index);
+          await wrapperAction.add(runFunction(creatorRecord) || {}, index);
           setLoading(false);
         }}
       >