+++ packages/layout/src/components/SiderMenu/BaseMenu.tsx
@@ -67,7 +67,7 @@ export type BaseMenuProps = {
         onClick: () => void;
       },
       defaultDom: React.ReactNode,
-      menuProps: BaseMenuProps,
+      menuProps: BaseMenuProps & Partial<PrivateSiderMenuProps>,
     ) => React.ReactNode
   >;